fix(attach): eliminate no-timeout wait chains (IMPL-22-HANG)
Session RCU pin + structural destroy, pool destroy-no-wait, notify
flush/shutdown split and watchdog fix, persist I/O worker with flush timeout;
add C-ATTACH-WATCH-CB-1 gate and sync Day20/21 concurrency docs.

+++ adapter/persistence/attach_runtime_sidecar.c
@@ -369,7 +369,7 @@ int bs_adapter_attach_persist_sidecar_validate(const char*          manifest_pat
             return 0;
         }
         char uri[4096];
-        if (uri_len + 1 > sizeof(uri))
+        if ((size_t)(uri_len + 1) > sizeof(uri))
         {
             free(data);
             return 0;
